+++ src/iipyper/__init__.py
@@ -9,6 +9,7 @@ from .timing import *
 from .midi import *
 from .osc import *
 from .oscmap import *
+from .update import *
 from .audio import *
 from .tui import *
 from .state import _lock
